Tilføjede funktioner igen, samt ændrede float i profile til double

+++ src/app/fifi.Core/LabeledValue.cs
@@ -8,13 +8,13 @@ namespace fifi.Core
 {
     public class LabeledValue
     {
-        public LabeledValue(string label, float value)
+        public LabeledValue(string label, double value)
         {
             Label = label;
             Value = value;
         }
 
         public string Label { get; private set; }
-        public float Value { get; private set; }
+        public double Value { get; private set; }
     }
 }
